+++ Application/EdFi.Ods.AdminApi/Infrastructure/Database/Commands/AddApplicationCommand.cs
@@ -105,7 +105,7 @@ public interface IAddApplicationModel
     int VendorId { get; }
     string? ClaimSetName { get; }
     IEnumerable<int>? ProfileIds { get; }
-    IEnumerable<int>? EducationOrganizationIds { get; }
+    IEnumerable<long>? EducationOrganizationIds { get; }
     IEnumerable<int>? OdsInstanceIds { get; }
 }
